Consolidate action form logic (#29326)
* Make `ActionForm` accept an `action` prop

* Make field settings inside `ActionForm` if missing

* Clean date-time values inside `ActionForm`

* WIP

* Clean and filter values on submit in `ActionForm`

* Sort `ActionVizForm` props

* Sort `Action` props

* Use more generic `WritebackAction` type

* Move more logic into the `ActionForm`

* Add `useActionForm` hook

* Fix parameter type in tests

* Fix submitting a form when all parameters mapped

* Move utils to `useActionForm`

* Fix submit button label in public action e2e tests

* Filter changed values only for implicit update actions

* Rename `cleanInitialValues` to `cleanedInitialValues`

* Fix `shouldShowConfirmation` can be a string

* Nice boring code

* Fix types in tests

* Small rename

* Move `useActionForm` utils closer to source (#29328)

* Fix `ActionForm` ignored submit errors

* Fix default values display in the form editor

* Fix names
