+++ frontend/src/metabase/actions/components/ActionForm/ActionForm.unit.spec.tsx
@@ -8,7 +8,10 @@ import type {
   ParametersForActionExecution,
   WritebackParameter,
 } from "metabase-types/api";
-import { createMockActionParameter } from "metabase-types/api/mocks";
+import {
+  createMockActionParameter,
+  createMockQueryAction,
+} from "metabase-types/api/mocks";
 
 import ActionForm from "./ActionForm";
 
@@ -43,22 +46,29 @@ type SetupOpts = {
   initialValues?: ParametersForActionExecution;
   parameters: WritebackParameter[];
   formSettings: ActionFormSettings;
+  onSubmit?: () => Promise<void>;
 };
 
-const setup = ({ initialValues, parameters, formSettings }: SetupOpts) => {
-  const onSubmit = jest.fn();
+const setup = ({
+  initialValues,
+  parameters,
+  formSettings,
+  onSubmit = jest.fn(),
+}: SetupOpts) => {
+  const action = createMockQueryAction({
+    parameters,
+    visualization_settings: formSettings,
+  });
 
   render(
     <ActionForm
+      action={action}
       initialValues={initialValues}
-      parameters={parameters}
-      formSettings={formSettings}
-      submitTitle="Save"
       onSubmit={onSubmit}
     />,
   );
 
-  return { onSubmit };
+  return { action, onSubmit };
 };
 
 describe("Actions > ActionForm", () => {
@@ -218,7 +228,7 @@ describe("Actions > ActionForm", () => {
     });
 
     it("can submit form field values", async () => {
-      const { onSubmit } = setup({
+      const { action, onSubmit } = setup({
         parameters: [
           makeParameter({ id: "abc-123" }),
           makeParameter({ id: "def-456" }),
@@ -242,7 +252,7 @@ describe("Actions > ActionForm", () => {
 
       userEvent.type(screen.getByLabelText(/text input/i), "Murloc");
       userEvent.type(screen.getByLabelText(/number input/i), "12345");
-      userEvent.click(screen.getByRole("button", { name: "Save" }));
+      userEvent.click(screen.getByRole("button", { name: action.name }));
 
       await waitFor(() => {
         expect(onSubmit).toHaveBeenCalledWith(
@@ -254,11 +264,47 @@ describe("Actions > ActionForm", () => {
         );
       });
     });
+
+    it("shows an error if submit fails", async () => {
+      const message = "Something went wrong when submitting the form.";
+      const error = { success: false, error: message, message };
+      const { action } = await setup({
+        onSubmit: jest.fn().mockRejectedValue(error),
+        parameters: [
+          makeParameter({ id: "abc-123" }),
+          makeParameter({ id: "def-456" }),
+        ],
+        formSettings: {
+          type: "form",
+          fields: {
+            "abc-123": makeFieldSettings({
+              inputType: "string",
+              id: "abc-123",
+              title: "text input",
+            }),
+            "def-456": makeFieldSettings({
+              inputType: "number",
+              id: "def-456",
+              title: "number input",
+            }),
+          },
+        },
+      });
+
+      userEvent.type(screen.getByLabelText(/text input/i), "Murloc");
+      userEvent.type(screen.getByLabelText(/number input/i), "12345");
+      userEvent.click(screen.getByRole("button", { name: action.name }));
+
+      expect(await screen.findByText(message)).toBeInTheDocument();
+      expect(
+        screen.getByRole("button", { name: action.name }),
+      ).toHaveTextContent("Failed");
+    });
   });
 
   describe("Form Validation", () => {
     it("allows form submission when required fields are provided", async () => {
-      const { onSubmit } = setup({
+      const { action, onSubmit } = setup({
         parameters: [
           makeParameter({ id: "abc-123" }),
           makeParameter({ id: "def-456" }),
@@ -284,14 +330,14 @@ describe("Actions > ActionForm", () => {
 
       userEvent.type(await screen.findByLabelText(/foo input/i), "baz");
       userEvent.type(await screen.findByLabelText(/bar input/i), "baz");
-      userEvent.click(screen.getByRole("button", { name: "Save" }));
+      userEvent.click(screen.getByRole("button", { name: action.name }));
 
       await waitFor(() => expect(onSubmit).toHaveBeenCalled());
       expect(screen.queryByText(/required/i)).not.toBeInTheDocument();
     });
 
     it("disables form submission when required fields are not provided", async () => {
-      const { onSubmit } = setup({
+      const { action, onSubmit } = setup({
         parameters: [
           makeParameter({ id: "abc-123" }),
           makeParameter({ id: "def-456" }),
@@ -318,17 +364,19 @@ describe("Actions > ActionForm", () => {
       userEvent.click(await screen.findByLabelText(/foo input/i)); // leave empty
       userEvent.type(await screen.findByLabelText(/bar input/i), "baz");
       await waitFor(() =>
-        expect(screen.getByRole("button", { name: "Save" })).toBeDisabled(),
+        expect(
+          screen.getByRole("button", { name: action.name }),
+        ).toBeDisabled(),
       );
 
-      userEvent.click(screen.getByRole("button", { name: "Save" }));
+      userEvent.click(screen.getByRole("button", { name: action.name }));
 
       expect(await screen.findByText(/required/i)).toBeInTheDocument();
       expect(onSubmit).not.toHaveBeenCalled();
     });
 
     it("allows form submission when all required fields are set", async () => {
-      const { onSubmit } = setup({
+      const { action, onSubmit } = setup({
         parameters: [
           makeParameter({ id: "abc-123" }),
           makeParameter({ id: "def-456" }),
@@ -352,21 +400,21 @@ describe("Actions > ActionForm", () => {
         },
       });
 
-      expect(screen.getByRole("button", { name: "Save" })).toBeDisabled();
+      expect(screen.getByRole("button", { name: action.name })).toBeDisabled();
 
       userEvent.type(screen.getByLabelText(/foo input/i), "baz");
       await waitFor(() => {
-        expect(screen.getByRole("button", { name: "Save" })).toBeEnabled();
+        expect(screen.getByRole("button", { name: action.name })).toBeEnabled();
       });
 
-      userEvent.click(screen.getByRole("button", { name: "Save" }));
+      userEvent.click(screen.getByRole("button", { name: action.name }));
       await waitFor(() => {
         expect(onSubmit).toHaveBeenCalled();
       });
     });
 
     it("allows form submission when all fields are optional", async () => {
-      const { onSubmit } = setup({
+      const { action, onSubmit } = setup({
         parameters: [
           makeParameter({ id: "abc-123" }),
           makeParameter({ id: "def-456" }),
@@ -390,9 +438,9 @@ describe("Actions > ActionForm", () => {
         },
       });
 
-      expect(screen.getByRole("button", { name: "Save" })).toBeEnabled();
+      expect(screen.getByRole("button", { name: action.name })).toBeEnabled();
 
-      userEvent.click(screen.getByRole("button", { name: "Save" }));
+      userEvent.click(screen.getByRole("button", { name: action.name }));
 
       await waitFor(() => expect(onSubmit).toHaveBeenCalled());
     });
@@ -415,7 +463,7 @@ describe("Actions > ActionForm", () => {
     });
 
     it("allows submission of a null non-required boolean field", async () => {
-      const { onSubmit } = setup({
+      const { action, onSubmit } = setup({
         parameters: [
           makeParameter({ id: "abc-123" }),
           makeParameter({ id: "def-456" }),
@@ -441,14 +489,14 @@ describe("Actions > ActionForm", () => {
 
       userEvent.type(await screen.findByLabelText(/foo input/i), "baz");
       userEvent.type(await screen.findByLabelText(/bar input/i), "baz");
-      userEvent.click(screen.getByRole("button", { name: "Save" }));
+      userEvent.click(screen.getByRole("button", { name: action.name }));
 
       await waitFor(() => expect(onSubmit).toHaveBeenCalled());
       expect(screen.queryByText(/required/i)).not.toBeInTheDocument();
     });
 
     it("sets a default value for an empty field", async () => {
-      const { onSubmit } = setup({
+      const { action, onSubmit } = setup({
         parameters: [
           makeParameter({ id: "abc-123" }),
           makeParameter({ id: "def-456" }),
@@ -474,14 +522,14 @@ describe("Actions > ActionForm", () => {
         },
       });
 
-      userEvent.click(screen.getByRole("button", { name: "Save" }));
+      userEvent.click(screen.getByRole("button", { name: action.name }));
 
       await waitFor(() => expect(onSubmit).toHaveBeenCalled());
       expect(screen.queryByText(/required/i)).not.toBeInTheDocument();
     });
 
     it("sets types on form submissions correctly", async () => {
-      const { onSubmit } = setup({
+      const { action, onSubmit } = setup({
         parameters: [
           makeParameter({ id: "abc-123" }),
           makeParameter({ id: "def-456" }),
@@ -515,7 +563,7 @@ describe("Actions > ActionForm", () => {
       userEvent.type(await screen.findByLabelText(/foo input/i), "1");
       userEvent.type(await screen.findByLabelText(/bar input/i), "1");
       userEvent.type(await screen.findByLabelText(/baz input/i), "1");
-      userEvent.click(screen.getByRole("button", { name: "Save" }));
+      userEvent.click(screen.getByRole("button", { name: action.name }));
 
       await waitFor(() => {
         expect(onSubmit).toHaveBeenCalledWith(
@@ -535,7 +583,7 @@ describe("Actions > ActionForm", () => {
     const inputTypes = ["string", "number", "text", "date", "datetime", "time"];
     inputTypes.forEach(inputType => {
       it(`casts empty optional ${inputType} field to null`, async () => {
-        const { onSubmit } = setup({
+        const { action, onSubmit } = setup({
           initialValues: { "abc-123": 1 },
           parameters: [makeParameter({ id: "abc-123" })],
           formSettings: {
@@ -555,7 +603,7 @@ describe("Actions > ActionForm", () => {
         fireEvent.change(screen.getByLabelText(/input/i), {
           target: { value: "" },
         });
-        userEvent.click(screen.getByRole("button", { name: "Save" }));
+        userEvent.click(screen.getByRole("button", { name: action.name }));
 
         await waitFor(() => {
           expect(onSubmit).toHaveBeenCalledWith(
@@ -571,7 +619,7 @@ describe("Actions > ActionForm", () => {
     // bug repro: https://github.com/metabase/metabase/issues/27377
     // eslint-disable-next-line jest/no-disabled-tests
     it.skip("casts empty optional category fields to null", async () => {
-      const { onSubmit } = setup({
+      const { action, onSubmit } = setup({
         initialValues: { "abc-123": "aaa" },
         parameters: [makeParameter({ id: "abc-123" })],
         formSettings: {
@@ -588,7 +636,7 @@ describe("Actions > ActionForm", () => {
       });
 
       userEvent.clear(screen.getByLabelText(/input/i));
-      userEvent.click(screen.getByRole("button", { name: "Save" }));
+      userEvent.click(screen.getByRole("button", { name: action.name }));
 
       await waitFor(() => {
         expect(onSubmit).toHaveBeenCalledWith(
