+++ frontend/src/metabase/actions/components/ActionViz/ActionVizForm.tsx
@@ -1,15 +1,16 @@
 import React, { useState } from "react";
 
+import useActionForm from "metabase/actions/hooks/use-action-form";
 import { getFormTitle } from "metabase/actions/utils";
 
 import type {
-  WritebackQueryAction,
-  WritebackParameter,
-  OnSubmitActionForm,
   ActionDashboardCard,
+  OnSubmitActionForm,
+  Dashboard,
   ParametersForActionExecution,
   VisualizationSettings,
-  Dashboard,
+  WritebackAction,
+  WritebackParameter,
 } from "metabase-types/api";
 
 import ActionParametersInputForm, {
@@ -21,32 +22,38 @@ import { shouldShowConfirmation } from "./utils";
 import { FormWrapper, FormTitle } from "./ActionForm.styled";
 
 interface ActionFormProps {
-  onSubmit: OnSubmitActionForm;
+  action: WritebackAction;
   dashcard: ActionDashboardCard;
+  dashboard: Dashboard;
+  missingParameters?: WritebackParameter[];
+  mappedParameters?: WritebackParameter[];
+  dashcardParamValues: ParametersForActionExecution;
   settings: VisualizationSettings;
   isSettings: boolean;
-  dashboard: Dashboard;
-  missingParameters: WritebackParameter[];
-  dashcardParamValues: ParametersForActionExecution;
-  action: WritebackQueryAction;
   shouldDisplayButton: boolean;
   isEditingDashcard: boolean;
+  onSubmit: OnSubmitActionForm;
 }
 
 function ActionVizForm({
-  onSubmit,
-  dashcard,
-  settings,
-  isSettings,
-  dashboard,
-  missingParameters,
-  dashcardParamValues,
   action,
+  dashcard,
+  dashboard,
+  settings,
+  missingParameters = [],
+  mappedParameters = [],
+  dashcardParamValues,
+  isSettings,
   shouldDisplayButton,
   isEditingDashcard,
+  onSubmit,
 }: ActionFormProps) {
   const [showModal, setShowModal] = useState(false);
   const title = getFormTitle(action);
+  const { getCleanValues } = useActionForm({
+    action,
+    initialValues: dashcardParamValues,
+  });
 
   // only show confirmation if there are no missing parameters
   const showConfirmMessage =
@@ -56,7 +63,7 @@ function ActionVizForm({
     if (missingParameters.length > 0 || showConfirmMessage) {
       setShowModal(true);
     } else {
-      onSubmit({});
+      onSubmit(getCleanValues());
     }
   };
 
@@ -72,24 +79,24 @@ function ActionVizForm({
     return (
       <>
         <ActionButtonView
-          onClick={onClick}
           settings={settings}
           isFullHeight={!isSettings}
           focus={isEditingDashcard}
+          onClick={onClick}
         />
         {showModal && (
           <ActionParametersInputModal
-            onClose={() => setShowModal(false)}
-            title={title}
-            onSubmit={onModalSubmit}
-            showConfirmMessage={!!showConfirmMessage}
-            confirmMessage={action.visualization_settings?.confirmMessage}
+            action={action}
             dashboard={dashboard}
             dashcard={dashcard}
-            missingParameters={missingParameters}
+            mappedParameters={mappedParameters}
             dashcardParamValues={dashcardParamValues}
+            title={title}
+            showConfirmMessage={showConfirmMessage}
+            confirmMessage={action.visualization_settings?.confirmMessage}
+            onSubmit={onModalSubmit}
+            onClose={() => setShowModal(false)}
             onCancel={() => setShowModal(false)}
-            action={action}
           />
         )}
       </>
@@ -100,12 +107,12 @@ function ActionVizForm({
     <FormWrapper>
       <FormTitle>{title}</FormTitle>
       <ActionParametersInputForm
-        onSubmit={onSubmit}
+        action={action}
         dashboard={dashboard}
         dashcard={dashcard}
-        missingParameters={missingParameters}
+        mappedParameters={mappedParameters}
         dashcardParamValues={dashcardParamValues}
-        action={action}
+        onSubmit={onSubmit}
       />
     </FormWrapper>
   );
