+++ frontend/src/metabase/actions/containers/ActionParametersInputForm/ActionParametersInputForm.tsx
@@ -1,16 +1,12 @@
 import React, { useCallback, useMemo, useState, useEffect } from "react";
 import { t } from "ttag";
+import _ from "underscore";
 
 import EmptyState from "metabase/components/EmptyState";
 
 import { ActionsApi, PublicApi } from "metabase/services";
 
 import ActionForm from "metabase/actions/components/ActionForm";
-import {
-  generateFieldSettingsFromParameters,
-  getSubmitButtonColor,
-  getSubmitButtonLabel,
-} from "metabase/actions/utils";
 import { getDashboardType } from "metabase/dashboard/utils";
 
 import type {
@@ -19,23 +15,18 @@ import type {
   Dashboard,
   ActionDashboardCard,
   ParametersForActionExecution,
-  ActionFormSettings,
   WritebackAction,
 } from "metabase-types/api";
-import type Field from "metabase-lib/metadata/Field";
-
-import { getChangedValues, getInitialValues } from "./utils";
 
 export interface ActionParametersInputFormProps {
   action: WritebackAction;
-  missingParameters?: WritebackParameter[];
-  dashcardParamValues?: ParametersForActionExecution;
   dashboard?: Dashboard;
   dashcard?: ActionDashboardCard;
-  onCancel?: () => void;
-  submitButtonColor?: string;
+  mappedParameters?: WritebackParameter[];
+  dashcardParamValues?: ParametersForActionExecution;
   onSubmit: OnSubmitActionForm;
   onSubmitSuccess?: () => void;
+  onCancel?: () => void;
 }
 
 const shouldPrefetchValues = (action: WritebackAction) =>
@@ -43,7 +34,7 @@ const shouldPrefetchValues = (action: WritebackAction) =>
 
 function ActionParametersInputForm({
   action,
-  missingParameters = action.parameters,
+  mappedParameters = [],
   dashcardParamValues = {},
   dashboard,
   dashcard,
@@ -51,37 +42,51 @@ function ActionParametersInputForm({
   onSubmit,
   onSubmitSuccess,
 }: ActionParametersInputFormProps) {
-  const [prefetchValues, setPrefetchValues] =
+  const [prefetchedValues, setPrefetchedValues] =
     useState<ParametersForActionExecution>({});
 
+  const hasPrefetchedValues = Object.keys(prefetchedValues).length > 0;
   const shouldPrefetch = useMemo(
     () => shouldPrefetchValues(action) && dashboard && dashcard,
     [action, dashboard, dashcard],
   );
 
-  const prefetchEndpoint =
-    getDashboardType(dashboard?.id) === "public"
-      ? PublicApi.prefetchValues
-      : ActionsApi.prefetchValues;
+  const initialValues = useMemo(
+    () => ({
+      ...prefetchedValues,
+      ...dashcardParamValues,
+    }),
+    [prefetchedValues, dashcardParamValues],
+  );
+
+  const hiddenFields = useMemo(
+    () => mappedParameters.map(parameter => parameter.id),
+    [mappedParameters],
+  );
 
   const fetchInitialValues = useCallback(async () => {
+    const prefetchEndpoint =
+      getDashboardType(dashboard?.id) === "public"
+        ? PublicApi.prefetchValues
+        : ActionsApi.prefetchValues;
+
     const fetchedValues = await prefetchEndpoint({
       dashboardId: dashboard?.id,
       dashcardId: dashcard?.id,
       parameters: JSON.stringify(dashcardParamValues),
-    }).catch(() => false);
+    }).catch(_.noop);
 
     if (fetchedValues) {
-      setPrefetchValues(fetchedValues);
+      setPrefetchedValues(fetchedValues);
     }
-  }, [dashboard?.id, dashcard?.id, dashcardParamValues, prefetchEndpoint]);
+  }, [dashboard?.id, dashcard?.id, dashcardParamValues]);
 
   useEffect(() => {
     const hasValueFromDashboard = Object.keys(dashcardParamValues).length > 0;
     const canPrefetch = hasValueFromDashboard && dashboard && dashcard;
 
     if (shouldPrefetch) {
-      setPrefetchValues({});
+      setPrefetchedValues({});
       canPrefetch && fetchInitialValues();
     }
   }, [
@@ -92,69 +97,32 @@ function ActionParametersInputForm({
     fetchInitialValues,
   ]);
 
-  const fieldSettings = useMemo(
-    () =>
-      action.visualization_settings?.fields ??
-      // if there are no field settings, we generate them from the parameters and field metadata
-      generateFieldSettingsFromParameters(
-        missingParameters,
-        dashcard?.card?.result_metadata as unknown as Field[],
-      ),
-    [action, missingParameters, dashcard],
-  );
-
-  const initialValues = useMemo(
-    () => getInitialValues(fieldSettings, prefetchValues),
-    [fieldSettings, prefetchValues],
-  );
-
   const handleSubmit = useCallback(
-    async (params, actions) => {
+    async (parameters, actions) => {
       actions.setSubmitting(true);
-      const paramsWithChangedValues = getChangedValues(params, initialValues);
-
-      const { success, error } = await onSubmit(paramsWithChangedValues);
-
+      const { success, error } = await onSubmit(parameters);
       if (success) {
         actions.setErrors({});
         onSubmitSuccess?.();
-
         shouldPrefetch ? fetchInitialValues() : actions.resetForm();
       } else {
         throw new Error(error);
       }
     },
-    [
-      onSubmit,
-      onSubmitSuccess,
-      initialValues,
-      fetchInitialValues,
-      shouldPrefetch,
-    ],
+    [shouldPrefetch, onSubmit, onSubmitSuccess, fetchInitialValues],
   );
-
-  const hasPrefetchedValues = !!Object.keys(prefetchValues).length;
 
   if (shouldPrefetch && !hasPrefetchedValues) {
     return <EmptyState message={t`Choose a record to update`} />;
   }
 
-  const submitButtonLabel = getSubmitButtonLabel(action);
-
-  const formSettings: ActionFormSettings = action.visualization_settings ?? {
-    type: "button",
-    fields: fieldSettings,
-  };
-
   return (
     <ActionForm
-      parameters={missingParameters}
-      formSettings={formSettings}
+      action={action}
       initialValues={initialValues}
-      onClose={onCancel}
+      hiddenFields={hiddenFields}
       onSubmit={handleSubmit}
-      submitTitle={submitButtonLabel}
-      submitButtonColor={getSubmitButtonColor(action)}
+      onClose={onCancel}
     />
   );
 }
