+++ frontend/src/metabase/actions/containers/ActionParametersInputForm/ActionParametersInputForm.unit.spec.tsx
@@ -7,37 +7,46 @@ import { waitFor } from "@testing-library/react";
 import { render, screen } from "__support__/ui";
 
 import {
+  createMockActionDashboardCard,
   createMockActionParameter,
   createMockQueryAction,
   createMockImplicitQueryAction,
   createMockDashboard,
 } from "metabase-types/api/mocks";
 
-import ActionParametersInputForm from "./ActionParametersInputForm";
-import ActionParametersInputModal from "./ActionParametersInputModal";
+import ActionParametersInputForm, {
+  ActionParametersInputFormProps,
+} from "./ActionParametersInputForm";
+import ActionParametersInputModal, {
+  ActionParametersInputModalProps,
+} from "./ActionParametersInputModal";
 
-const mockAction = createMockQueryAction();
-const defaultProps = {
-  missingParameters: [
-    createMockActionParameter({
-      id: "parameter_1",
-      type: "type/Text",
-    }),
-    createMockActionParameter({
-      id: "parameter_2",
-      type: "type/Text",
-    }),
-  ],
-  dashcardParamValues: {},
+const parameter1 = createMockActionParameter({
+  id: "parameter_1",
+  type: "type/Text",
+});
+
+const parameter2 = createMockActionParameter({
+  id: "parameter_2",
+  type: "type/Text",
+});
+
+const mockAction = createMockQueryAction({
+  parameters: [parameter1, parameter2],
+});
+
+const defaultProps: ActionParametersInputFormProps = {
   action: mockAction,
+  mappedParameters: [],
   dashboard: createMockDashboard({ id: 123 }),
-  dashcard: createMockDashboard({ id: 456 }),
-  onSubmit: jest.fn(() => ({ success: true })),
+  dashcard: createMockActionDashboardCard({ id: 456, action: mockAction }),
+  dashcardParamValues: {},
   onCancel: _.noop,
   onSubmitSuccess: _.noop,
+  onSubmit: jest.fn().mockResolvedValue({ success: true }),
 };
 
-async function setup(options?: any) {
+function setup(options?: Partial<ActionParametersInputModalProps>) {
   render(<ActionParametersInputForm {...defaultProps} {...options} />);
 }
 
@@ -75,7 +84,7 @@ describe("Actions > ActionParametersInputForm", () => {
   });
 
   it("passes form values to submit handler", async () => {
-    const submitSpy = jest.fn(() => ({ success: true }));
+    const submitSpy = jest.fn().mockResolvedValue({ success: true });
     await setup({
       onSubmit: submitSpy,
     });
@@ -101,17 +110,19 @@ describe("Actions > ActionParametersInputForm", () => {
   });
 
   it("should generate field types from parameter types", async () => {
-    const missingParameters = [
-      createMockActionParameter({
-        id: "parameter_1",
-        type: "type/Text",
-      }),
-      createMockActionParameter({
-        id: "parameter_2",
-        type: "type/Integer",
-      }),
-    ];
-    await setup({ missingParameters });
+    const action = createMockQueryAction({
+      parameters: [
+        createMockActionParameter({
+          id: "parameter_1",
+          type: "type/Text",
+        }),
+        createMockActionParameter({
+          id: "parameter_2",
+          type: "type/Integer",
+        }),
+      ],
+    });
+    await setup({ action });
 
     expect(screen.getByPlaceholderText("Parameter 1")).toHaveAttribute(
       "type",
@@ -126,11 +137,15 @@ describe("Actions > ActionParametersInputForm", () => {
   it("should fetch and load existing values from API for implicit update actions", async () => {
     setupPrefetch();
 
+    const idParameter = createMockActionParameter({ id: "id" });
+
     await setup({
       action: createMockImplicitQueryAction({
         type: "implicit",
         kind: "row/update",
+        parameters: [idParameter, parameter1, parameter2],
       }),
+      mappedParameters: [idParameter],
       dashcardParamValues: {
         id: 888,
       },
@@ -163,7 +178,6 @@ describe("Actions > ActionParametersInputForm", () => {
         type: "implicit",
         kind: "row/delete",
       }),
-      missingParameters: [],
       showConfirmMessage: true,
     });
