+++ frontend/src/metabase/actions/containers/ActionParametersInputForm/ActionParametersInputModal.tsx
@@ -8,19 +8,22 @@ import ActionParametersInputForm, {
 } from "./ActionParametersInputForm";
 
 interface ModalProps {
-  onClose: () => void;
   title: string;
   showConfirmMessage?: boolean;
   confirmMessage?: string;
+  onClose: () => void;
 }
 
-export default function ActionParametersInputModal({
-  onClose,
+export type ActionParametersInputModalProps = ModalProps &
+  ActionParametersInputFormProps;
+
+function ActionParametersInputModal({
   title,
   showConfirmMessage,
   confirmMessage,
+  onClose,
   ...formProps
-}: ModalProps & ActionParametersInputFormProps) {
+}: ActionParametersInputModalProps) {
   return (
     <Modal onClose={onClose}>
       <ModalContent title={title} onClose={onClose}>
@@ -36,3 +39,5 @@ export default function ActionParametersInputModal({
 const ConfirmMessage = ({ message }: { message?: string }) => (
   <div>{message ?? t`This action cannot be undone.`}</div>
 );
+
+export default ActionParametersInputModal;
