+++ frontend/src/metabase/actions/hooks/use-action-form/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./use-action-form";
